requirements.txt
# AI/LLM Framework
openai>=1.0.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Agent Framework (可选择其一)
# langgraph>=0.0.20
# autogen>=0.2.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
loguru>=0.7.0

# Vector DB (for RAG)
chromadb>=0.4.0
# faiss-cpu>=1.7.4

# Report Generation
jinja2>=3.1.0
reportlab>=4.0.0  # PDF generation
markdown>=3.5.0
